requirements.txt
Package                  Version
------------------------ -----------
beautifulsoup4           4.11.1
Brotli                   1.0.9
bs4                      0.0.1
cachetools               5.1.0
certifi                  2022.5.18.1
charset-normalizer       2.0.12
google-api-core          2.8.0
google-api-python-client 2.48.0
google-auth              2.6.6
google-auth-httplib2     0.1.0
googleapis-common-protos 1.56.1
httplib2                 0.20.4
idna                     3.3
isort                    5.10.1
mutagen                  1.45.1
pip                      22.1.1
protobuf                 3.20.1
pyasn1                   0.4.8
pyasn1-modules           0.2.8
pycryptodomex            3.14.1
pyparsing                3.0.9
python-dateutil          2.8.2
requests                 2.27.1
rsa                      4.8
setuptools               62.3.2
six                      1.16.0
soupsieve                2.3.2.post1
uritemplate              4.1.1
urllib3                  1.26.9
websockets               10.3
yt-dlp                   2022.5.18
